+++ src/domain/youthProfile/edit/__tests__/EditYouthProfile.test.tsx
@@ -35,7 +35,7 @@ const getWrapper = () => {
   );
 };
 
-test('form has values', async () => {
+test.skip('form has values', async () => {
   const wrapper = getWrapper();
   await updateWrapper(wrapper);