+++ CHANGELOG.md
@@ -8,4 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Form fields are now validated on blur
 - Profile creation and editing views are now divided into sections [#123](https://github.com/City-of-Helsinki/youth-membership-ui/pull/123)
 
+### Added
+- Controls for additional contact persons
+
 ## [1.0.0-rc.3]